requirements.txt
Django==3.0.2
psycopg2==2.8.4
djangorestframework==3.11.0
django-cors-headers==3.2.1
pyjwt==1.7.1
django-extensions==2.2.6